requirements.txt
click
progress
pyyaml
simplejson
# Pin to cryptography==3.4.8 since the cryptography-3.4.x branch is the
# last branch that can be built without the Rust toolchain.
# More info on issue: https://github.com/canonical/hotsos/issues/326
# [TODO] Bump cryptography to more recent release
cryptography==3.4.8
searchkit
structr
fasteners
